Adding Progressbar to EventCreate
Goal:
We want to start the progress bar before the createEvent() action and stop the progress bar if there's any error. We don't want to stop the progress bar until it successfully loads the eventShow page, after we've made the event and load the new show component for that event.

// EventCreate.js
// import NProgress library
// Inside createEvent() we start progress bar: NProgress.start()
// Inside .catch, we end the progress bar.

That's it!