Creating 404 Error Pages
Problem: navigating to page that doesn't exist renders a blank page (with navbar component)

Goal:
1. Implement catch all page by creating a NotFound.vue component
2. Add component to our router.js
3. Add a catch-all route to our router.js

// create new component NotFound.vue inside /views
// Add template and router-link to go back to the homepage (event listings)
// import component into router.js
// Add path for /404
// Add catch all path that will redirect any path's that do not match any coded paths and load up the 404-notFound component.

Works but with new issue.

Problem: when navigating to an event that doesn't exist (/event/###) it loads an error notification regarding event fetch. We don't have a 404 catch for an event that doesn't exist.

Solution:
1. Remove the error catch from the Event module
2. Catch the error in our router, and redirect
3. Add parameter to 404 route and component