Using Vuelidate
Created example component (example.vue)

Using vuelidate, we added an email input inside a form that checks if the email is valid, makes sure it is required, displays error messages if there are errors when focus has entered then left the input field using $touch() and @blur and we submit the field when user submits and log to console.

We also disabled the submit button if any errors exist from vuelidate.