Added Login, helpers
Added a getter called loggedIn that returns a truthy value based on if a user is logged in using state.user

Logged in, like when registering, sends credentials to the server and if there's a match, the server responds with a new token which we save to axios headers and redirect to the dashboard.

User login and register buttons and pages were also added.